+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- [bind] also create links for inline_components dependencies in the components directory
+
 ## [0.6.3] - 2017-05-21
 
 - fix to generation of links for inline-components